glm/requirements.txt
pyserial==3.5
numpy==2.2.0
asyncio==3.4.3
websockets==15.0.0